+++ CHANGELOG.md
@@ -92,6 +92,12 @@ Also made the random-positions (used when exiting a theme) slightly more 'themat
 Also fixed a bug where search did not work when a student was already selected.
 
 #### W03: Production
+
+- [ ] banner
+- [ ] banner, integrate with aarch.dk
+- [ ] banner, cta
+- [ ] banner, fallback
+
 #### W04: Release